로그인 여부/규칙 DB저장

+++ back/accounts/api.py
@@ -11,6 +11,9 @@ from ninja.errors import HttpError
 import logging
 import os
 from django.contrib.auth.hashers import make_password #해시화
+from django.contrib.auth.decorators import login_required
+from django.http import JsonResponse
+import json
 router = Router()
 logger = logging.getLogger(__name__)
 
@@ -28,7 +31,7 @@ class CreateUserSchema(Schema):
     MBTI: str
     position: str
     blog: Optional[HttpUrl]
-    individual_rule: str
+    individual_rule: list
     # profile_image:
     
 #로그인/로그아웃   
@@ -99,9 +102,10 @@ def create_user(request, data: CreateUserSchema):
         position = data.position,
         github = data.github,
         blog = data.blog,
-        individual_rule = data.individual_rule,
+        individual_rule = str(data.individual_rule)
         # profile_image:,
     )
+    # print(str(user.individual_rule))
     user.save()
     user_data = {
         "id": user.id,
@@ -114,17 +118,17 @@ def create_user(request, data: CreateUserSchema):
         "position": user.position,
         "github": user.github,
         "blog": user.blog,
-        "individual_rule": user.individual_rule,
+        "individual_rule": user.individual_rule
         # Add other fields as needed
     }
     return user_data
 
 #회원조회
+@login_required
 @router.get("/get-user/{user_id}",tags=["회원가입"])
 def get_user(request, user_id: int):
     try:
         user = User.objects.get(id=user_id)
-
         serialized_user = {
             "id":user.id,
             "username": user.username,
@@ -139,14 +143,20 @@ def get_user(request, user_id: int):
             "individual_rule": user.individual_rule,
             # profile_image:,
         }
+        # json_str = user.individual_rule.replace("'", "\"")
+        # data_list = json.loads(json_str)
+        # for i in data_list:
+        #     print(user.id)
+        #     print(dict(i).get("fee"))
+        #     print(dict(i).get("rule"))
         return serialized_user
     except User.DoesNotExist:
         return {"message": "실패"}, 404
-    
+
 #회원수정
+@login_required
 @router.put("/update-user/{user_id}", tags=["회원가입"])
 def update_user(request, user_id: int, data: CreateUserSchema):
-    
     try:
         user = User.objects.get(id=user_id)
         if data.password:#수정시 비밀번호 해시화
@@ -178,27 +188,29 @@ def update_user(request, user_id: int, data: CreateUserSchema):
         "github": user.github,
         "blog": user.blog,
         "individual_rule": user.individual_rule,
-         # user.profile_image=data
+        # user.profile_image=data
         }
         return user_data
     except User.DoesNotExist:
         return {"message": "실패"}, 404
     
-#회원탈퇴    
+#회원탈퇴
+@login_required
 @router.delete("/delete-user/{user_id}", tags=["회원가입"])
 def delete_user(request, user_id: int):
     try:
         user = User.objects.get(id=user_id)
-        data_user={
-            "user_id":user.id
-        }
-        user.delete()
+        if user.id:
+            data_user={
+                "user_id":user.id
+            }
+            user.delete()
         return data_user
     except User.DoesNotExist:
         return {"message": "실패"}, 404
 
 #로그인
-@router.post("/login", tags=["로그인/로그아웃"])
+@router.post("/login", tags=["로그인/로그아웃/여부"])
 def login_user(request, data: LoginInput):
     user = authenticate(request, username=data.username, password=data.password)
     if user is not None:
@@ -211,7 +223,8 @@ def login_user(request, data: LoginInput):
         raise HttpError(401, "실패")
     
 #로그아웃
-@router.post("/logout", tags=["로그인/로그아웃"])
+@login_required
+@router.post("/logout", tags=["로그인/로그아웃/여부"])
 def logout_user(request):
     if request.user.is_authenticated:
         user={
@@ -221,3 +234,26 @@ def logout_user(request):
         return user
     else:
         raise HttpError(401, "실패")
+    
+@login_required#아니면 로그인
+@router.get("/status", tags=["로그인/로그아웃/여부"])
+def check_login_status(request):#로그인 여부
+    if request.user.is_authenticated:
+        user = User.objects.get(id=request.user.id)
+        serialized_user = {
+            "id":user.id,
+            "username": user.username,
+            "first_name":user.first_name,
+            "phone": user.phone,
+            "address": user.address,
+            "github": user.github,
+            "blog": user.blog,
+            "MBTI": user.MBTI,
+            "position": user.position,
+            "birth": user.birth.strftime('%Y-%m-%d'),
+            "individual_rule": user.individual_rule,
+            # profile_image:,
+        }
+        return serialized_user
+    else:#확인
+        return JsonResponse({'is_logged_in': False})
